@@ -5,7 +5,7 @@ homepage: https://github.com/corsac-dart/jwt
 author: Anatoly Pulyaevskiy <anatoly.pulyaevskiy@gmail.com>
 
 environment:
-  sdk: ">=2.11.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   asn1lib: ^1.0.0
@@ -16,6 +16,6 @@ dependencies:
 
 dev_dependencies:
   pedantic: ^1.11.0
-  test: ^1.16.8
-  test_coverage:
-    git: https://github.com/priezz/test-coverage
+  test: ^1.17.4
+  # test_coverage:
+    # git: https://github.com/priezz/test-coverage
